+++ konva.js
@@ -8,7 +8,7 @@
    * Konva JavaScript Framework v7.0.6
    * http://konvajs.org/
    * Licensed under the MIT
-   * Date: Fri Aug 21 2020
+   * Date: Wed Aug 26 2020
    *
    * Original work Copyright (C) 2011 - 2013 by Eric Rowell (KineticJS)
    * Modified work Copyright (C) 2014 - present by Anton Lavrenov (Konva)
@@ -1283,6 +1283,21 @@
                       ' is a not valid value for "' +
                       attr +
                       '" attribute. The value should be a string.');
+              }
+              return val;
+          };
+      }
+  }
+  function getStringOrGradientValidator() {
+      if (Konva.isUnminified) {
+          return function (val, attr) {
+              var isString = Util._isString(val);
+              var isGradient = Object.prototype.toString.call(val) === '[object CanvasGradient]';
+              if (!(isString || isGradient)) {
+                  Util.warn(_formatValue(val) +
+                      ' is a not valid value for "' +
+                      attr +
+                      '" attribute. The value should be a string or a native gradient.');
               }
               return val;
           };
@@ -7379,7 +7394,7 @@
   Shape.prototype.on.call(Shape.prototype, 'fillPriorityChange.konva fillLinearGradientColorStopsChange.konva fillLinearGradientStartPointXChange.konva fillLinearGradientStartPointYChange.konva fillLinearGradientEndPointXChange.konva fillLinearGradientEndPointYChange.konva', _clearLinearGradientCache);
   Shape.prototype.on.call(Shape.prototype, 'fillPriorityChange.konva fillRadialGradientColorStopsChange.konva fillRadialGradientStartPointXChange.konva fillRadialGradientStartPointYChange.konva fillRadialGradientEndPointXChange.konva fillRadialGradientEndPointYChange.konva fillRadialGradientStartRadiusChange.konva fillRadialGradientEndRadiusChange.konva', _clearRadialGradientCache);
   // add getters and setters
-  Factory.addGetterSetter(Shape, 'stroke', undefined, getStringValidator());
+  Factory.addGetterSetter(Shape, 'stroke', undefined, getStringOrGradientValidator());
   /**
    * get/set stroke color
    * @name Konva.Shape#stroke
@@ -7697,7 +7712,7 @@
    * };
    * imageObj.src = 'path/to/image/jpg';
    */
-  Factory.addGetterSetter(Shape, 'fill', undefined, getStringValidator());
+  Factory.addGetterSetter(Shape, 'fill', undefined, getStringOrGradientValidator());
   /**
    * get/set fill color
    * @name Konva.Shape#fill
@@ -18258,8 +18273,8 @@
           RGBA: RGBA,
           Sepia: Sepia,
           Solarize: Solarize,
-          Threshold: Threshold
-      }
+          Threshold: Threshold,
+      },
   });
 
   // main entry for umd build for rollup